Make Rectangle and Point generic classes
Now we can express that some parameters need to be Rectangle<int> or
Point<int> objects. There are some minor breaking changes which
hopefully will not affect many developers.

+++ lib/src/engine/render_state.dart
@@ -78,7 +78,7 @@ class RenderState {
 
     cs2.matrix.copyFromAndConcat(matrix, cs1.matrix);
     cs2.compositeOperation = (composite is String) ? composite : cs1.compositeOperation;
-    cs2.alpha = alpha * cs1.alpha.toDouble();
+    cs2.alpha = alpha * cs1.alpha;
 
     _currentContextState = cs2;
     displayObject._renderInternal(this);
